requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
asgiref==3.4.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
clang==5.0
click==8.0.1
colorama==0.4.4
discord.py==1.7.3
flatbuffers==1.12
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
image==1.5.33
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
mathparse==0.1.2
multidict==5.1.0
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
Pillow==8.3.1
Pint==0.17
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.0.1
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.7.5
pytz==2021.1
PyYAML==3.13
regex==2021.8.3
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3

